requirements.txt
colorama==0.4.6
Flask==3.1.0
oletools==0.60.2
pefile==2024.8.26
psutil==6.1.0
pyssdeep==1.0.0
PyYAML==6.0.2
Werkzeug==3.1.3
